move invalid error to a different file
